// product/src/modules/grpc/proto/v2.pb.ts
export interface Category {
  id: string;
  type: string;
  name?: string | undefined;
  nameAr?: string | undefined;
}
// ...
export interface Variant {
  id: string;
  name: string;
  nameAr: string;
}
// ...
export interface GetViewedProductsRequest {
  productIds: string[];
  shouldSkipExpire?: boolean | undefined;
  categoryId?: string | undefined;
  getSpecificCategory?: boolean | undefined;
}
// ...
export interface Title {
  arName: string;
  enName: string;
}

export interface Value {
  arName: string;
  enName: string;
}

export interface Attribute {
  title: Title | undefined;
  value: Value | undefined;
  iconURL?: string | undefined;
}

export interface GetViewedProductsResponse {
  products: GetViewedProductsResponse_Product[];
}

export interface GetViewedProductsResponse_Product {
  id: string;
  grade: string;
  gradeAr: string;
  deviceModel: DeviceModel | undefined;
  variant: Variant | undefined;
  attributes: Attribute[];
  isGreatDeal: boolean;
  isMerchant: boolean;
  originalPrice: number;
  productImage: string;
  productImages: string[];
  sellPrice: number;
  tags: string;
  sellStatus: string;
  sellDate: string;
  condition: Condition | undefined;
  showSecurityBadge: boolean;
  brand: Category | undefined;
  category: Category | undefined;
  grandTotal: number;
  vat: number;
  shippingCharge: number;
  buyAmount: number;
  expressDeliveryBadge?: boolean | undefined;
  year?: string | undefined;
}

export interface DeviceModel {
  name: string;
  nameAr: string;
}
// ...
export interface Condition {
  id: string;
  name: string;
  nameAr: string;
  labelColor: string;
  textColor: string;
}